+++ Android/UdacityExercises/Lesson08-Quiz-Example/T08.03-Exercise-FinishQuizExample/app/src/main/java/com/udacity/example/quizexample/MainActivity.java
@@ -23,6 +23,7 @@ import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
+import android.widget.TextView;
 
 import com.udacity.example.droidtermsprovider.DroidTermsExampleContract;
 
@@ -37,6 +38,12 @@ public class MainActivity extends AppCompatActivity {
 
     // The current state of the app
     private int mCurrentState;
+
+    private int mDefCol;
+    private int mWordCol;
+
+    private TextView mWordTextView;
+    private TextView mDefinitionTextView;
 
     private Button mButton;
 
@@ -55,12 +62,12 @@ public class MainActivity extends AppCompatActivity {
         setContentView(R.layout.activity_main);
 
         // Get the views
-        // TODO (1) You'll probably want more than just the Button
         mButton = (Button) findViewById(R.id.button_next);
+        mWordTextView = (TextView) findViewById(R.id.text_view_word);
+        mDefinitionTextView = (TextView) findViewById(R.id.text_view_definition);
 
         //Run the database operation to get the cursor off of the main thread
         new WordFetchTask().execute();
-
     }
 
     /**
@@ -83,31 +90,34 @@ public class MainActivity extends AppCompatActivity {
     }
 
     public void nextWord() {
+        if (mData != null) {
+            if (!mData.moveToNext())
+                mData.moveToFirst();
 
-        // Change button text
-        mButton.setText(getString(R.string.show_definition));
+            mDefinitionTextView.setVisibility(View.INVISIBLE);
+            mButton.setText(getString(R.string.show_definition));
 
-        // TODO (3) Go to the next word in the Cursor, show the next word and hide the definition
-        // Note that you shouldn't try to do this if the cursor hasn't been set yet.
-        // If you reach the end of the list of words, you should start at the beginning again.
-        mCurrentState = STATE_HIDDEN;
+            mWordTextView.setText(mData.getString(mWordCol));
+            mDefinitionTextView.setText(mData.getString(mDefCol));
 
+            mCurrentState = STATE_HIDDEN;
+        }
     }
 
     public void showDefinition() {
+        if (mData != null) {
+            mDefinitionTextView.setVisibility(View.VISIBLE);
 
-        // Change button text
-        mButton.setText(getString(R.string.next_word));
+            mButton.setText(getString(R.string.next_word));
 
-        // TODO (4) Show the definition
-        mCurrentState = STATE_SHOWN;
-
+            mCurrentState = STATE_SHOWN;
+        }
     }
 
     @Override
     protected void onDestroy() {
         super.onDestroy();
-        // TODO (5) Remember to close your cursor!
+        mData.close();
     }
 
     // Use an async task to do the data fetch off of the main thread.
@@ -136,8 +146,10 @@ public class MainActivity extends AppCompatActivity {
             // Set the data for MainActivity
             mData = cursor;
 
-            // TODO (2) Initialize anything that you need the cursor for, such as setting up
-            // the screen with the first word and setting any other instance variables
+            mDefCol = mData.getColumnIndex(DroidTermsExampleContract.COLUMN_DEFINITION);
+            mWordCol = mData.getColumnIndex(DroidTermsExampleContract.COLUMN_WORD);
+
+            nextWord();
         }
     }